# crypto-trading-connector-be/internal/generated/models.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package generated

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for BalanceCurrency.
const (
	JPY BalanceCurrency = "JPY"
)

// Defines values for ChartResponsePeriod.
const (
	ChartResponsePeriodAll  ChartResponsePeriod = "all"
	ChartResponsePeriodN1y  ChartResponsePeriod = "1y"
	ChartResponsePeriodN24h ChartResponsePeriod = "24h"
	ChartResponsePeriodN30d ChartResponsePeriod = "30d"
	ChartResponsePeriodN7d  ChartResponsePeriod = "7d"
)

// Defines values for CreateOrderRequestOrderType.
const (
	CreateOrderRequestOrderTypeLimit CreateOrderRequestOrderType = "limit"
)

// Defines values for CreateOrderRequestPair.
const (
	CreateOrderRequestPairBTCJPY CreateOrderRequestPair = "BTC/JPY"
	CreateOrderRequestPairETHJPY CreateOrderRequestPair = "ETH/JPY"
)

// Defines values for ErrorResponseError.
const (
	BADREQUEST          ErrorResponseError = "BAD_REQUEST"
	INSUFFICIENTBALANCE ErrorResponseError = "INSUFFICIENT_BALANCE"
	INTERNALERROR       ErrorResponseError = "INTERNAL_ERROR"
	INTERNALSERVERERROR ErrorResponseError = "INTERNAL_SERVER_ERROR"
	INVALIDAMOUNT       ErrorResponseError = "INVALID_AMOUNT"
	INVALIDPRICE        ErrorResponseError = "INVALID_PRICE"
	INVALIDREQUEST      ErrorResponseError = "INVALID_REQUEST"
	NOTFOUND            ErrorResponseError = "NOT_FOUND"
	UNAUTHORIZED        ErrorResponseError = "UNAUTHORIZED"
	UNSUPPORTEDPAIR     ErrorResponseError = "UNSUPPORTED_PAIR"
)

// Defines values for OrderOrderType.
const (
	OrderOrderTypeLimit OrderOrderType = "limit"
)

// Defines values for OrderPair.
const (
	OrderPairBTCJPY OrderPair = "BTC/JPY"
	OrderPairETHJPY OrderPair = "ETH/JPY"
)

// Defines values for OrderStatus.
const (
	Cancelled OrderStatus = "cancelled"
	Completed OrderStatus = "completed"
	Failed    OrderStatus = "failed"
	Pending   OrderStatus = "pending"
)

// Defines values for GetCryptoChartParamsPeriod.
const (
	GetCryptoChartParamsPeriodAll  GetCryptoChartParamsPeriod = "all"
	GetCryptoChartParamsPeriodN1y  GetCryptoChartParamsPeriod = "1y"
	GetCryptoChartParamsPeriodN24h GetCryptoChartParamsPeriod = "24h"
	GetCryptoChartParamsPeriodN30d GetCryptoChartParamsPeriod = "30d"
	GetCryptoChartParamsPeriodN7d  GetCryptoChartParamsPeriod = "7d"
)

// Balance defines model for Balance.
type Balance struct {
	// AvailableBalance Available balance in JPY
	AvailableBalance float64 `json:"availableBalance"`

	// Currency Currency code
	Currency BalanceCurrency `json:"currency"`

	// Timestamp Unix timestamp of the balance snapshot
	Timestamp int64 `json:"timestamp"`
}

// BalanceCurrency Currency code
type BalanceCurrency string

// ChartDataPoint defines model for ChartDataPoint.
type ChartDataPoint struct {
	// Day Day label (Mon, Tue, Wed, etc.)
	Day string `json:"day"`

	// Price Price at this data point
	Price float64 `json:"price"`
}

// ChartResponse defines model for ChartResponse.
type ChartResponse struct {
	// Data Array of chart data points
	Data []ChartDataPoint `json:"data"`

	// Period Time period of the chart data
	Period ChartResponsePeriod `json:"period"`
}

// ChartResponsePeriod Time period of the chart data
type ChartResponsePeriod string

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	// Amount Amount of cryptocurrency to buy
	Amount float64 `json:"amount"`

	// OrderType Order type (currently only limit orders supported)
	OrderType CreateOrderRequestOrderType `json:"orderType"`

	// Pair Trading pair
	Pair CreateOrderRequestPair `json:"pair"`

	// Price Limit price in JPY
	Price float64 `json:"price"`
}

// CreateOrderRequestOrderType Order type (currently only limit orders supported)
type CreateOrderRequestOrderType string

// CreateOrderRequestPair Trading pair
type CreateOrderRequestPair string

// CryptoData defines model for CryptoData.
type CryptoData struct {
	// ChangePercent Percentage change (positive or negative)
	ChangePercent float64 `json:"changePercent"`

	// ChartData Array of chart data points
	ChartData []ChartDataPoint `json:"chartData"`

	// CurrentPrice Current price in JPY
	CurrentPrice float64 `json:"currentPrice"`

	// Icon Icon character or emoji
	Icon string `json:"icon"`

	// IconColor Hex color code for the icon background
	IconColor string `json:"iconColor"`

	// Id Unique identifier for the cryptocurrency
	Id string `json:"id"`

	// Name Full name of the cryptocurrency
	Name string `json:"name"`

	// Pair Trading pair
	Pair string `json:"pair"`

	// Symbol Trading symbol
	Symbol string `json:"symbol"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Details Additional error details (optional)
	Details *map[string]interface{} `json:"details,omitempty"`

	// Error Error type
	Error ErrorResponseError `json:"error"`

	// Message Human-readable error message
	Message string `json:"message"`
}

// ErrorResponseError Error type
type ErrorResponseError string

// MarketResponse defines model for MarketResponse.
type MarketResponse struct {
	// Data Array of cryptocurrency market data
	Data []CryptoData `json:"data"`

	// Timestamp Unix timestamp of the response
	Timestamp int64 `json:"timestamp"`
}

// Order defines model for Order.
type Order struct {
	// Amount Amount of cryptocurrency
	Amount float64 `json:"amount"`

	// CreatedAt Order creation timestamp
	CreatedAt time.Time `json:"createdAt"`

	// EstimatedTotal Estimated total cost in JPY (price * amount)
	EstimatedTotal float64 `json:"estimatedTotal"`

	// OrderId Unique order identifier
	OrderId openapi_types.UUID `json:"orderId"`

	// OrderType Order type
	OrderType OrderOrderType `json:"orderType"`

	// Pair Trading pair
	Pair OrderPair `json:"pair"`

	// Price Limit price in JPY
	Price float64 `json:"price"`

	// Status Order status
	Status OrderStatus `json:"status"`

	// UpdatedAt Order last update timestamp
	UpdatedAt *time.Time `json:"updatedAt,omitempty"`
}

// OrderOrderType Order type
type OrderOrderType string

// OrderPair Trading pair
type OrderPair string

// OrderStatus Order status
type OrderStatus string

// GetCryptoChartParams defines parameters for GetCryptoChart.
type GetCryptoChartParams struct {
	// Period Time period for chart data
	Period *GetCryptoChartParamsPeriod `form:"period,omitempty" json:"period,omitempty"`
}

// GetCryptoChartParamsPeriod defines parameters for GetCryptoChart.
type GetCryptoChartParamsPeriod string

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrderRequest