# openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Trading Connector API
  description: REST API for cryptocurrency trading platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@crypto-trading-connector.com

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: crypto
    description: Cryptocurrency market data operations
  - name: orders
    description: Order management operations
  - name: balance
    description: Balance and wallet operations

paths:
  /crypto/market:
    get:
      tags:
        - crypto
      summary: Get all cryptocurrency market data
      description: Returns a list of all cryptocurrencies with their current market data including prices, changes, and chart data
      operationId: getCryptoMarket
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /crypto/{id}:
    get:
      tags:
        - crypto
      summary: Get cryptocurrency by ID
      description: Returns detailed information about a specific cryptocurrency
      operationId: getCryptoById
      parameters:
        - name: id
          in: path
          required: true
          description: Cryptocurrency ID (e.g., bitcoin, ethereum)
          schema:
            type: string
            example: bitcoin
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CryptoData'
        '404':
          description: Cryptocurrency not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /crypto/{id}/chart:
    get:
      tags:
        - crypto
      summary: Get chart data for cryptocurrency
      description: Returns historical price chart data for a specific cryptocurrency
      operationId: getCryptoChart
      parameters:
        - name: id
          in: path
          required: true
          description: Cryptocurrency ID
          schema:
            type: string
            example: bitcoin
        - name: period
          in: query
          required: false
          description: Time period for chart data
          schema:
            type: string
            enum: [24h, 7d, 30d, 1y, all]
            default: 7d
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartResponse'
        '404':
          description: Cryptocurrency not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders:
    post:
      tags:
        - orders
      summary: Create a new order
      description: Creates a new limit order for cryptocurrency trading
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /balance:
    get:
      tags:
        - balance
      summary: Get user balance
      description: Returns the current available balance for the authenticated user
      operationId: getBalance
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CryptoData:
      type: object
      required:
        - id
        - name
        - symbol
        - pair
        - icon
        - iconColor
        - currentPrice
        - changePercent
        - chartData
      properties:
        id:
          type: string
          description: Unique identifier for the cryptocurrency
          example: bitcoin
        name:
          type: string
          description: Full name of the cryptocurrency
          example: Bitcoin
        symbol:
          type: string
          description: Trading symbol
          example: BTC
        pair:
          type: string
          description: Trading pair
          example: BTC/JPY
        icon:
          type: string
          description: Icon character or emoji
          example: ₿
        iconColor:
          type: string
          description: Hex color code for the icon background
          example: "#f7931a"
        currentPrice:
          type: number
          format: double
          description: Current price in JPY
          example: 9850000
        changePercent:
          type: number
          format: double
          description: Percentage change (positive or negative)
          example: 5.2
        chartData:
          type: array
          description: Array of chart data points
          items:
            $ref: '#/components/schemas/ChartDataPoint'

    ChartDataPoint:
      type: object
      required:
        - day
        - price
      properties:
        day:
          type: string
          description: Day label (Mon, Tue, Wed, etc.)
          example: Mon
        price:
          type: number
          format: double
          description: Price at this data point
          example: 9350000

    MarketResponse:
      type: object
      required:
        - data
        - timestamp
      properties:
        data:
          type: array
          description: Array of cryptocurrency market data
          items:
            $ref: '#/components/schemas/CryptoData'
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the response
          example: 1704067200

    ChartResponse:
      type: object
      required:
        - data
        - period
      properties:
        data:
          type: array
          description: Array of chart data points
          items:
            $ref: '#/components/schemas/ChartDataPoint'
        period:
          type: string
          description: Time period of the chart data
          enum: [24h, 7d, 30d, 1y, all]
          example: 7d

    CreateOrderRequest:
      type: object
      required:
        - pair
        - orderType
        - price
        - amount
      properties:
        pair:
          type: string
          description: Trading pair
          enum: [BTC/JPY, ETH/JPY]
          example: BTC/JPY
        orderType:
          type: string
          description: Order type (currently only limit orders supported)
          enum: [limit]
          example: limit
        price:
          type: number
          format: double
          description: Limit price in JPY
          minimum: 0
          example: 14000000
        amount:
          type: number
          format: double
          description: Amount of cryptocurrency to buy
          minimum: 0.001
          example: 0.001

    Order:
      type: object
      required:
        - orderId
        - pair
        - orderType
        - price
        - amount
        - estimatedTotal
        - status
        - createdAt
      properties:
        orderId:
          type: string
          format: uuid
          description: Unique order identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        pair:
          type: string
          description: Trading pair
          enum: [BTC/JPY, ETH/JPY]
          example: BTC/JPY
        orderType:
          type: string
          description: Order type
          enum: [limit]
          example: limit
        price:
          type: number
          format: double
          description: Limit price in JPY
          example: 14000000
        amount:
          type: number
          format: double
          description: Amount of cryptocurrency
          example: 0.001
        estimatedTotal:
          type: number
          format: double
          description: Estimated total cost in JPY (price * amount)
          example: 14000
        status:
          type: string
          description: Order status
          enum: [pending, completed, cancelled, failed]
          example: pending
        createdAt:
          type: string
          format: date-time
          description: Order creation timestamp
          example: 2024-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          description: Order last update timestamp
          example: 2024-01-01T00:00:00Z

    Balance:
      type: object
      required:
        - availableBalance
        - currency
        - timestamp
      properties:
        availableBalance:
          type: number
          format: double
          description: Available balance in JPY
          example: 1540200
        currency:
          type: string
          description: Currency code
          enum: [JPY]
          example: JPY
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the balance snapshot
          example: 1704067200

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          enum: 
            - NOT_FOUND
            - BAD_REQUEST
            - INVALID_REQUEST
            - INVALID_AMOUNT
            - INVALID_PRICE
            - UNSUPPORTED_PAIR
            - INSUFFICIENT_BALANCE
            - UNAUTHORIZED
            - INTERNAL_ERROR
            - INTERNAL_SERVER_ERROR
          example: INSUFFICIENT_BALANCE
        message:
          type: string
          description: Human-readable error message
          example: Available balance is insufficient for this order
        details:
          type: object
          description: Additional error details (optional)
          additionalProperties: true